+++ src/Common/AbstractBuilder.php
@@ -38,6 +38,42 @@ abstract class AbstractBuilder
 
     /**
      *
+     * Builds the WITH clause.
+     *
+     * @param array $with The CTE elements.
+     *
+     * @param bool $recursive True if recursive, false if not.
+     *
+     * @return string
+     *
+     */
+    public function buildWith(array $with, $recursive = false)
+    {
+        if (empty($with)) {
+            return ''; // not applicable
+        }
+
+        $keyword = $recursive && $this->allowsRecursiveKeyword()
+            ? 'WITH RECURSIVE'
+            : 'WITH';
+
+        return $keyword . ' ' . implode(',' . PHP_EOL, $with) . PHP_EOL;
+    }
+
+    /**
+     *
+     * Does this dialect allow the RECURSIVE keyword in WITH?
+     *
+     * @return bool
+     *
+     */
+    protected function allowsRecursiveKeyword()
+    {
+        return true;
+    }
+
+    /**
+     *
      * Builds the `WHERE` clause of the statement.
      *
      * @param array $where The WHERE elements.